update package and refactoring react quill
