+++ src/components/Common/CustomEditor.jsx
@@ -4,15 +4,137 @@ import ReactQuill from "react-quill";
 import { CiEdit } from "react-icons/ci";
 import {
   convertToPersian,
+  SanitizeHTML,
   getFieldTranslationByNames,
 } from "../../services/Utility";
-import {
-  EditorContainer as StyledEditorContainer,
-  Label,
-  Char,
-  formats,
-  modulesWithoutImage,
-} from "../editorContainerStyle";
+import styled from "styled-components";
+
+const EditorContainer = styled.div`
+  background-color: ${(props) =>
+    props.theme.colors.newColors.otherColors.inputBg};
+  border-radius: 5px;
+  overflow: hidden;
+  color: white;
+  margin: 10px auto;
+  height: 212px;
+  border: ${({ border }) => (border ? "1px solid gray" : "none")};
+  .ql-toolbar {
+    background-color: ${(props) =>
+      props.theme.colors.newColors.otherColors.inputBg};
+    border: none;
+    border-bottom: 1px solid gray;
+  }
+
+  .ql-container {
+    background-color: ${(props) =>
+      props.theme.colors.newColors.otherColors.inputBg};
+    color: ${(props) => props.theme.colors.newColors.shades.title};
+    border: none;
+    overflow: auto;
+    max-height: 150px;
+  }
+
+  && .ql-editor {
+    min-height: 150px;
+    text-align: unset;
+    font-size: 18px !important;
+    line-height: 1.6;
+    -webkit-text-size-adjust: 100%;
+    font-family: "AzarMehr" !important;
+  }
+
+  /* placeholder */
+  && .ql-editor::before {
+    font-size: inherit !important;
+    color: #888;
+    opacity: 0.7;
+    font-family: "AzarMehr" !important;
+  }
+
+  .ql-toolbar .ql-picker {
+    color: white;
+  }
+
+  .ql-toolbar .ql-stroke {
+    stroke: ${(props) => props.theme.colors.newColors.shades.title};
+  }
+
+  .ql-toolbar .ql-fill {
+    fill: ${(props) => props.theme.colors.newColors.shades.title};
+  }
+
+  .ql-toolbar .ql-picker-options {
+    border: 1px solid #555;
+  }
+
+  @media (max-width: 700px) {
+    && .ql-editor {
+      font-size: 15px !important;
+      -webkit-text-size-adjust: 100%;
+    }
+  }
+`;
+
+const Label = styled.h2`
+  color: ${(props) => props.theme.colors.newColors.shades.title};
+  display: block;
+  margin-bottom: 10px;
+  font-weight: 500;
+  font-size: 16px;
+  margin-top: 20px;
+`;
+
+const Char = styled.div`
+  display: flex;
+  justify-content: end;
+  align-items: center;
+  gap: 5px;
+  svg {
+    color: ${({ isOverLimit, theme }) =>
+      isOverLimit ? "red" : theme.colors.newColors.shades.title};
+  }
+
+  span {
+    color: ${({ isOverLimit }) => (isOverLimit ? "red" : "#a0a0ab")};
+    font-size: 13px;
+    font-weight: 400;
+  }
+`;
+const formats = [
+  "size",
+  "bold",
+  "italic",
+  "underline",
+  "strike",
+  "blockquote",
+  "list",
+  "bullet",
+  "indent",
+  "link",
+  "code-block",
+  "align",
+];
+
+const getModules = (img = false) => {
+  const toolbar = [
+    ["bold", "italic", "underline", "strike", "blockquote"],
+    [
+      { list: "ordered" },
+      { list: "bullet" },
+      { indent: "-1" },
+      { indent: "+1" },
+    ],
+    ["link", "code-block"], // دکمه تصویر به صورت داینامیک اضافه می‌شود
+    [{ align: [] }],
+  ];
+
+  if (img) {
+    // اضافه کردن دکمه image در صورت فعال بودن پراپ
+    toolbar[2].splice(1, 0, "image"); // دکمه image بعد از link اضافه می‌شود
+  }
+
+  return { toolbar };
+};
 
 /**
  * Reusable RichTextEditor with strict char limit
@@ -33,6 +155,7 @@ const CustomEditor = ({
   showIcon = true,
   placeholder = "",
   border = false,
+  img = false,
 }) => {
   const [content, setContent] = useState(value);
 
@@ -40,21 +163,20 @@ const CustomEditor = ({
     setContent(value);
   }, [value]);
   const handleChange = (val, delta, source, editor) => {
-    // طول واقعی متن بدون html
-    const text = editor.getText(); // این متن plain text است
+    const text = editor.getText();
     let newValue = val;
 
     if (text.length - 1 > charLimit) {
-      // -1 چون editor یه \n اضافه می‌کنه
       const allowedText = text.slice(0, charLimit);
-      // جایگزینی متن editor با محدودیت
       const quill = editor;
-      quill.deleteText(charLimit, text.length); // بقیه رو حذف می‌کنه
+      quill.deleteText(charLimit, text.length);
       newValue = quill.root.innerHTML;
     }
 
-    setContent(newValue);
-    onChange?.(newValue);
+    const safeValue = SanitizeHTML(newValue);
+
+    setContent(safeValue);
+    onChange?.(safeValue);
   };
 
   const handleKeyDown = (event) => {
@@ -70,16 +192,22 @@ const CustomEditor = ({
       event.preventDefault();
     }
   };
-
   const handlePaste = (event) => {
     event.preventDefault();
+
     const paste = event.clipboardData.getData("text");
     const remaining = charLimit - content.length;
-    if (remaining <= 0) return; // هیچ چیزی اضافه نشود
-    const toPaste = paste.slice(0, remaining);
-    const newValue = content + toPaste;
-    setContent(newValue);
-    onChange?.(newValue);
+
+    if (paste.length > remaining) {
+      return;
+    }
+
+    const newValue = content + paste;
+
+    const safeValue = SanitizeHTML(newValue);
+
+    setContent(safeValue);
+    onChange?.(safeValue);
   };
 
   const remainingChars = charLimit - content.length;
@@ -88,17 +216,17 @@ const CustomEditor = ({
   return (
     <div>
       {label && <Label>{label}</Label>}
-      <StyledEditorContainer border={border}>
+      <EditorContainer border={border}>
         <ReactQuill
           value={content}
           onChange={handleChange}
           onKeyDown={handleKeyDown}
           onPaste={handlePaste}
-          modules={modulesWithoutImage}
+          modules={getModules(img)}
           formats={formats}
           placeholder={placeholder}
         />
-      </StyledEditorContainer>
+      </EditorContainer>
       <Char isOverLimit={isOverLimit}>
         <span>
           {convertToPersian(remainingChars)} {getFieldTranslationByNames("530")}
