--- src/components/editorContainerStyle.jsx
@@ -1,134 +1,0 @@
-import styled from "styled-components";
-export const EditorContainer = styled.div`
-  background-color: ${(props) =>
-    props.theme.colors.newColors.otherColors.inputBg};
-  border-radius: 5px;
-  overflow: hidden;
-  color: white;
-  margin: 10px auto;
-  height: 212px;
-  border: ${({ border }) => (border ? "1px solid gray" : "none")}; 
-  .ql-toolbar {
-    background-color: ${(props) =>
-      props.theme.colors.newColors.otherColors.inputBg};
-    border: none;
-    border-bottom: 1px solid gray;
-  }
-
-  .ql-container {
-    background-color: ${(props) =>
-      props.theme.colors.newColors.otherColors.inputBg};
-    color: ${(props) => props.theme.colors.newColors.shades.title};
-    border: none;
-    overflow: auto;
-    max-height: 150px;
-  }
-
-  && .ql-editor {
-    min-height: 150px;
-    text-align: unset;
-    font-size: 18px !important;
-    line-height: 1.6;
-    -webkit-text-size-adjust: 100%;
-    font-family: "AzarMehr" !important;
-  }
-
-  /* placeholder */
-  && .ql-editor::before {
-    font-size: inherit !important;
-    color: #888;
-    opacity: 0.7;
-    font-family: "AzarMehr" !important;
-  }
-
-  .ql-toolbar .ql-picker {
-    color: white;
-  }
-
-  .ql-toolbar .ql-stroke {
-    stroke: ${(props) => props.theme.colors.newColors.shades.title};
-  }
-
-  .ql-toolbar .ql-fill {
-    fill: ${(props) => props.theme.colors.newColors.shades.title};
-  }
-
-  .ql-toolbar .ql-picker-options {
-    border: 1px solid #555;
-  }
-
-  @media (max-width: 700px) {
-    && .ql-editor {
-      font-size: 15px !important;
-      -webkit-text-size-adjust: 100%;
-    }
-  }
-`;
-
-export const Label = styled.h2`
-  color: ${(props) => props.theme.colors.newColors.shades.title};
-  display: block;
-  margin-bottom: 10px;
-  font-weight: 500;
-  font-size: 16px;
-  margin-top: 20px;
-`;
-
-export const Char = styled.div`
-  display: flex;
-  justify-content: end;
-  align-items: center;
-  gap: 5px;
-  svg {
-    color: ${({ isOverLimit, theme }) =>
-      isOverLimit ? "red" : theme.colors.newColors.shades.title};
-  }
-
-  span {
-    color: ${({ isOverLimit }) => (isOverLimit ? "red" : "#a0a0ab")};
-    font-size: 13px;
-    font-weight: 400;
-  }
-`;
-export const formats = [
-  "size",
-  "bold",
-  "italic",
-  "underline",
-  "strike",
-  "blockquote",
-  "list",
-  "bullet",
-  "indent",
-  "link",
-  "code-block",
-  "align",
-];
-export const modules = {
-  toolbar: [
-    ["bold", "italic", "underline", "strike", "blockquote"],
-    [
-      { list: "ordered" },
-      { list: "bullet" },
-      { indent: "-1" },
-      { indent: "+1" },
-    ],
-    ["link", "image", "code-block"],
-    [{ align: [] }],
-  ],
-};
-
-//توجه شود که با بالایی فرق دارد برای استفاده در کامپوننت های گزارش  خطا
-export const modulesWithoutImage = {
-  toolbar: [
-    ["bold", "italic", "underline", "strike", "blockquote"],
-    [
-      { list: "ordered" },
-      { list: "bullet" },
-      { indent: "-1" },
-      { indent: "+1" },
-    ],
-    ["link", "code-block"],
-    [{ align: [] }],
-  ],
-};
