+++ src/pages/Sanad/Tabs/SendFiles.jsx
@@ -80,40 +80,28 @@ const SendFiles = ({ files, onFilesChange }) => {
   const [error, setError] = useState("");
   const fileInputRef = useRef(null);
   const MAX_FILE_SIZE_MB = 9;
+
   useEffect(() => {
-    if (!files || (Array.isArray(files) && files.length === 0)) {
+    if (files && files.length > 0) {
+      const file = files[0];
+      const previewUrl = file.type.startsWith("image/")
+        ? URL.createObjectURL(file)
+        : nonPhoto;
+      setPreview(previewUrl);
+    } else {
       setPreview("");
-      return;
     }
-
-    let objectUrl = "";
-
-    if (files instanceof File) {
-      if (files.type.startsWith("image/")) {
-        objectUrl = URL.createObjectURL(files);
-        setPreview(objectUrl);
-      } else {
-        setPreview(nonPhoto);
-      }
-    } else if (typeof files === "string") {
-      const isImage = /\.(jpg|jpeg|png|gif)$/i.test(files);
-      setPreview(isImage ? files : nonPhoto);
-    }
-
-    return () => {
-      if (objectUrl) URL.revokeObjectURL(objectUrl);
-    };
   }, [files]);
 
   const fileHandler = (e) => {
     setError("");
-    const newFile = e.target.files[0];
 
+    const newFile = e.target.files[0];
     if (newFile) {
       if (newFile.size > MAX_FILE_SIZE_MB * 1024 * 1024) {
-        setError(getFieldTranslationByNames("1482"));
+        setError(getFieldTranslationByNames("1643"));
       } else {
-        onFilesChange(newFile);
+        onFilesChange([newFile]);
       }
     }
   };
