+++ src/pages/Sanad/Tabs/notes/WriteNoteInput.jsx
@@ -1,41 +1,14 @@
-import "react-quill/dist/quill.snow.css";
-
-import { CiEdit } from "react-icons/ci";
-import ReactQuill from "react-quill";
-import { EditorContainer,Char,Label,formats,modules} from "../../../../components/editorContainerStyle";
-
-import {
-  convertToPersian,
-  getFieldTranslationByNames,
-} from "../../../../services/Utility";
+import CustomEditor from "../../../../components/Common/CustomEditor";
+import { getFieldTranslationByNames } from "../../../../services/Utility";
 
 const WriteNoteInput = ({ description, onChange }) => {
-  const charLimit = 2000;
-  const currentLength = description.length;
-  const remainingChars = charLimit - currentLength;
-  const isOverLimit = remainingChars <= 0;
-
-
   return (
-    <>
-      <Label>{getFieldTranslationByNames("1360")}</Label>
-      <EditorContainer>
-        <ReactQuill
-          value={description}
-          onChange={onChange}
-          modules={modules}
-          formats={formats}
-          // placeholder="یادداشت خود را بنویسید"
-        />
-      </EditorContainer>
-      <Char isOverLimit={isOverLimit}>
-        <span>
-          {convertToPersian(remainingChars)}{" "}
-          {getFieldTranslationByNames("530")}
-        </span>
-        <CiEdit size={20} />
-      </Char>
-    </>
+    <CustomEditor
+      value={description}
+      onChange={onChange}
+      label={getFieldTranslationByNames("1360")}
+      img
+    />
   );
 };
