--- src/pages/Sanad/Tabs/sent/ReplyInput.jsx
@@ -1,173 +1,0 @@
-import "react-quill/dist/quill.snow.css";
-
-import { CiEdit } from "react-icons/ci";
-import ReactQuill from "react-quill";
-
-import styled from "styled-components";
-import { useGlobalState } from "../GlobalVodStateProvider";
-import { useState } from "react";
-import { convertToPersian } from "../../../../services/Utility";
-
-const EditorContainer = styled.div`
-  background-color: #2c2c2c;
-  border-radius: 5px;
-  overflow: hidden;
-  color: white;
-
-  margin: 10px auto;
-  height: 212px;
-  overflow: auto;
-
-  .ql-toolbar {
-    background-color: #2c2c2c;
-    border: none;
-    border-bottom: 1px solid gray;
-  }
-
-  .ql-container {
-    background-color: #2c2c2c;
-    color: #606060;
-    font-family: inherit;
-    border: none;
-
-    text-align: unset;
-  }
-
-  .ql-editor {
-    min-height: 150px;
-
-    text-align: unset;
-  }
-
-  .ql-editor::before {
-    content: attr(data-placeholder);
-    color: #a0a0ab;
-    font-style: italic;
-    position: absolute;
-    left: 0;
-    right: 20px;
-    font-family: inherit;
-    text-align: unset;
-    pointer-events: none;
-    display: block;
-  }
-
-  .ql-toolbar .ql-picker {
-    color: white;
-  }
-
-  .ql-toolbar .ql-stroke {
-    stroke: white;
-  }
-
-  .ql-toolbar .ql-fill {
-    fill: white;
-  }
-
-  .ql-toolbar .ql-picker-label,
-  .ql-toolbar .ql-picker-options {
-    color: white;
-    background-color: #444;
-  }
-
-  .ql-toolbar .ql-picker-options {
-    border: 1px solid #555;
-  }
-
-  .ql-toolbar .ql-picker-label:hover,
-  .ql-toolbar .ql-picker-options:hover {
-    color: #ddd;
-    background-color: #555;
-  }
-`;
-
-const Char = styled.div`
-  display: flex;
-  justify-content: end;
-  align-items: center;
-  gap: 5px;
-
-  svg {
-    color: ${({ isOverLimit }) => (isOverLimit ? "red" : "#ffffff")};
-  }
-
-  span {
-    color: ${({ isOverLimit }) => (isOverLimit ? "red" : "#a0a0ab")};
-    font-size: 13px;
-    font-weight: 400;
-  }
-`;
-
-const Label = styled.h2`
-  color: #ffffff;
-  display: block;
-  margin-bottom: 10px;
-  font-weight: 500;
-  font-size: 16px;
-  margin-top: 20px;
-`;
-
-const ReplyInput = () => {
-  const [replyText, setReplyText] = useState("");
-  const charLimit = 2000;
-
-  const handleChange = (value) => {
-    if (value.length <= charLimit) {
-      setReplyText(value);
-    }
-  };
-
-  const currentLength = replyText.length;
-  const remainingChars = charLimit - currentLength;
-  const isOverLimit = remainingChars <= 0;
-
-  const modules = {
-    toolbar: [
-      ["bold", "italic", "underline", "strike", "blockquote"],
-      [
-        { list: "ordered" },
-        { list: "bullet" },
-        { indent: "-1" },
-        { indent: "+1" },
-      ],
-      ["link", "code-block"],
-      [{ align: [] }],
-    ],
-  };
-
-  const formats = [
-    "size",
-    "bold",
-    "italic",
-    "underline",
-    "strike",
-    "blockquote",
-    "list",
-    "bullet",
-    "indent",
-    "link",
-    "code-block",
-    "align",
-  ];
-
-  return (
-    <>
-      <Label>پاسخ دادن به سند</Label>
-      <EditorContainer>
-        <ReactQuill
-          value={replyText}
-          onChange={handleChange}
-          modules={modules}
-          formats={formats}
-          // placeholder="پاسخ خود را بنویسید"
-        />
-      </EditorContainer>
-      <Char isOverLimit={isOverLimit}>
-        <span>{convertToPersian(remainingChars)} کاراکتر</span>
-        <CiEdit size={20} />
-      </Char>
-    </>
-  );
-};
-
-export default ReplyInput;
