+++ src/pages/Settings/Tabs/aboutme-tab/NextYear.jsx
@@ -1,11 +1,8 @@
-import "react-quill/dist/quill.snow.css";
-import { CiEdit } from "react-icons/ci";
-import ReactQuill from "react-quill";
 import { convertToPersian } from "../../../../services/Utility";
 import { useGlobalState } from "./aboutGlobalStateProvider";
 import { getFieldTranslationByNames } from "../../../../services/Utility";
+import CustomEditor from "../../../../components/Common/CustomEditor";
 import { useState, useEffect } from "react";
-import { EditorContainer, Label, Char, formats, modules } from "../../../../components/editorContainerStyle";
 import { useLanguage } from "../../../../services/reducers/LanguageContext";
 import moment from "moment-jalaali";
 
@@ -35,40 +32,20 @@ const NextYear = () => {
     }
   };
 
-  const handleKeyDown = (event) => {
-    if (predictionValue.length >= charLimit && event.key !== "Backspace" && event.key !== "Delete") {
-      event.preventDefault();
-    }
-  };
-
-  const currentLength = predictionValue.length;
-  const remainingChars = charLimit - currentLength;
-  const isOverLimit = remainingChars <= 0;
-  const localizedRemainingChars = convertToPersian(remainingChars);
-
   const yearToDisplay = isPersian ? nextJalaliYear : nextGeorgianYear;
 
   return (
-    <>
-      <Label style={{ paddingTop: 20 }}>
-        {getFieldTranslationByNames("656")} {convertToPersian(yearToDisplay)}
-      </Label>
-
-      <EditorContainer>
-        <ReactQuill
-          value={predictionValue}
-          onChange={handleChange}
-          onKeyDown={handleKeyDown}
-          modules={modules}
-          formats={formats}
-        />
-      </EditorContainer>
-
-      <Char isOverLimit={isOverLimit}>
-        <span>{localizedRemainingChars} {getFieldTranslationByNames("530")}</span>
-        <CiEdit size={20} />
-      </Char>
-    </>
+    <div style={{ paddingTop: 20 }}>
+      <CustomEditor
+        label={
+          getFieldTranslationByNames("656") + convertToPersian(yearToDisplay)
+        }
+        value={predictionValue}
+        onChange={handleChange}
+        charLimit={10000}
+        img
+      />
+    </div>
   );
 };
